+++ Plugins/Player/Player.cpp
@@ -93,6 +93,7 @@ Player::Player(Services::ProxyServiceList* services)
     REGISTER(SetCustomToken);
     REGISTER(SetCreatureNameOverride);
     REGISTER(FloatingTextStringOnCreature);
+    REGISTER(ToggleDM);
 
 #undef REGISTER
 
@@ -126,14 +127,9 @@ ArgumentStack Player::ForcePlaceableExamineWindow(ArgumentStack&& args)
     {
         const auto placeableId = Services::Events::ExtractArgument<ObjectID>(args);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerExamineGui_PlaceableData(pPlayer, placeableId);
-        }
-        else
-        {
-            LOG_ERROR("Unable to get CNWSMessage");
         }
     }
 
@@ -195,14 +191,9 @@ ArgumentStack Player::StartGuiTimingBar(ArgumentStack&& args)
           ASSERT_OR_THROW(type > 0);
           ASSERT_OR_THROW(type <= 10);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerGuiTimingEvent(pPlayer, true, type, milliseconds);
-        }
-        else
-        {
-            LOG_ERROR("Unable to get CNWSMessage");
         }
     }
 
@@ -213,14 +204,9 @@ ArgumentStack Player::StopGuiTimingBar(ArgumentStack&& args)
 {
     if (auto *pPlayer = player(args))
     {
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerGuiTimingEvent(pPlayer, false, 10, 0);
-        }
-        else
-        {
-            LOG_ERROR("Unable to get CNWSMessage");
         }
     }
 
@@ -293,11 +279,13 @@ ArgumentStack Player::GetQuickBarSlot(ArgumentStack&& args)
         auto slot = Services::Events::ExtractArgument<int32_t>(args);
           ASSERT_OR_THROW(slot < 36);
 
-        CNWSCreature *pCreature = Globals::AppManager()->m_pServerExoApp->GetCreatureByGameObjectID(pPlayer->m_oidNWSObject);
-        if (!pCreature->m_pQuickbarButton)
-            pCreature->InitializeQuickbar();
+        if (auto *pCreature = Globals::AppManager()->m_pServerExoApp->GetCreatureByGameObjectID(pPlayer->m_oidNWSObject))
+        {
+            if (!pCreature->m_pQuickbarButton)
+                pCreature->InitializeQuickbar();
 
-        qbs = pCreature->m_pQuickbarButton[slot];
+            qbs = pCreature->m_pQuickbarButton[slot];
+        }
     }
 
     return Services::Events::Arguments
@@ -344,7 +332,7 @@ ArgumentStack Player::SetQuickBarSlot(ArgumentStack&& args)
         pCreature->m_pQuickbarButton[slot].m_oidSecondaryItem = Services::Events::ExtractArgument<ObjectID>(args);
         pCreature->m_pQuickbarButton[slot].m_oidItem          = Services::Events::ExtractArgument<ObjectID>(args);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
+        auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage();
         pMessage->SendServerToPlayerGuiQuickbar_SetButton(pPlayer, slot, 0);
     }
     return Services::Events::Arguments();
@@ -372,8 +360,7 @@ ArgumentStack Player::ShowVisualEffect(ArgumentStack&& args)
         pos.y = Services::Events::ExtractArgument<float>(args);
         pos.x = Services::Events::ExtractArgument<float>(args);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             ObjectVisualTransformData ovtd;
             pMessage->SendServerToPlayerArea_VisualEffect(pPlayer, effectId, pos, ovtd);
@@ -393,8 +380,7 @@ ArgumentStack Player::ChangeBackgroundMusic(ArgumentStack&& args)
           ASSERT_OR_THROW(track >= 0);
           ASSERT_OR_THROW(track <= 0xFFFF);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerAmbientMusicChangeTrack(oidPlayer, day, track);
         }
@@ -410,8 +396,7 @@ ArgumentStack Player::PlayBackgroundMusic(ArgumentStack&& args)
 
         auto play = Services::Events::ExtractArgument<int32_t>(args);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerAmbientMusicPlay(oidPlayer, play);
         }
@@ -429,8 +414,7 @@ ArgumentStack Player::ChangeBattleMusic(ArgumentStack&& args)
           ASSERT_OR_THROW(track >= 0);
           ASSERT_OR_THROW(track <= 0xFFFF);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerAmbientBattleMusicChange(oidPlayer, track);
         }
@@ -446,8 +430,7 @@ ArgumentStack Player::PlayBattleMusic(ArgumentStack&& args)
 
         auto play = Services::Events::ExtractArgument<int32_t>(args);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerAmbientBattleMusicPlay(oidPlayer, play);
         }
@@ -470,8 +453,7 @@ ArgumentStack Player::PlaySound(ArgumentStack&& args)
             oidTarget = pPlayer->m_oidNWSObject;
         }
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerAIActionPlaySound(playerID, oidTarget, sound.c_str());
         }
@@ -486,8 +468,7 @@ ArgumentStack Player::SetPlaceableUsable(ArgumentStack&& args)
         const auto oidPlaceable = Services::Events::ExtractArgument<ObjectID>(args);
         const auto bUsable = Services::Events::ExtractArgument<int32_t>(args);
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->CreateWriteMessage(sizeof(bUsable) + sizeof(oidPlaceable), pPlayer->m_nPlayerID, 1);
 
@@ -570,8 +551,7 @@ ArgumentStack Player::ApplyInstantVisualEffectToObject(ArgumentStack&& args)
         vTargetPosition.y = 0.0f;
         vTargetPosition.z = 0.0f;
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerGameObjUpdateVisEffect(
                     pPlayer,
@@ -595,8 +575,7 @@ ArgumentStack Player::UpdateCharacterSheet(ArgumentStack&& args)
         uint32_t msg = charSheet->ComputeCharacterSheetUpdateRequired(pPlayer);
         if (msg)
         {
-            auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-            if (pMessage)
+            if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
                 pMessage->WriteGameObjUpdate_CharacterSheet(pPlayer, msg);
         }
     }
@@ -614,7 +593,7 @@ ArgumentStack Player::OpenInventory(ArgumentStack&& args)
         CNWSPlayerInventoryGUI *pInventory = pPlayer->m_oidNWSObject == oidTarget ? pPlayer->m_pInventoryGUI :
                                                                                     pPlayer->m_pOtherInventoryGUI;
 
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
+        auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage();
         if (pMessage && pInventory)
         {
             pMessage->SendPlayerToServerGuiInventory_Status(pPlayer, open, oidTarget);
@@ -632,7 +611,7 @@ ArgumentStack Player::OpenInventory(ArgumentStack&& args)
 
 ArgumentStack Player::GetAreaExplorationState(ArgumentStack&& args)
 {
-    std::string encString = "";
+    std::string encString;
 
     if (auto *pPlayer = player(args))
     {
@@ -1138,7 +1117,7 @@ ArgumentStack Player::UpdateItemName(ArgumentStack&& args)
           ASSERT_OR_THROW(oidItem != Constants::OBJECT_INVALID);
 
         auto *pItem = Utils::AsNWSItem(Utils::GetGameObject(oidItem));
-        auto *pMessage = static_cast<CNWSMessage*>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
+        auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage();
 
         if (pItem && pMessage)
         {
@@ -1329,14 +1308,9 @@ ArgumentStack Player::SetResManOverride(ArgumentStack&& args)
         const auto newResName = Services::Events::ExtractArgument<std::string>(args);
           ASSERT_OR_THROW(newResName.size() <= 16);
 
-        auto *pMessage = static_cast<CNWSMessage *>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerResmanOverride(pPlayer->m_nPlayerID, resType, oldResName.c_str(), newResName.c_str());
-        }
-        else
-        {
-            LOG_ERROR("Unable to get CNWSMessage");
         }
     }
 
@@ -1350,14 +1324,9 @@ ArgumentStack Player::SetCustomToken(ArgumentStack&& args)
         const auto tokenNumber = Services::Events::ExtractArgument<int32_t>(args);
         const auto tokenText = Services::Events::ExtractArgument<std::string>(args);
 
-        auto *pMessage = static_cast<CNWSMessage *>(Globals::AppManager()->m_pServerExoApp->GetNWSMessage());
-        if (pMessage)
+        if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
         {
             pMessage->SendServerToPlayerSetCustomToken(pPlayer->m_nPlayerID, tokenNumber, tokenText.c_str());
-        }
-        else
-        {
-            LOG_ERROR("Unable to get CNWSMessage");
         }
     }
 
@@ -1450,6 +1419,44 @@ ArgumentStack Player::FloatingTextStringOnCreature(ArgumentStack&& args)
                 messageData.SetString(0, text);
 
                 pMessage->SendServerToPlayerCCMessage(pPlayer->m_nPlayerID, Constants::MessageClientSideMsgMinor::Feedback, &messageData, nullptr);
+            }
+        }
+    }
+
+    return Services::Events::Arguments();
+}
+
+ArgumentStack Player::ToggleDM(ArgumentStack&& args)
+{
+    if (auto *pPlayer = player(args))
+    {
+        auto isDM = !!Services::Events::ExtractArgument<int32_t>(args);
+        auto *pNetLayer = Globals::AppManager()->m_pServerExoApp->GetNetLayer();
+
+        if (auto *pPlayerInfo = pNetLayer->GetPlayerInfo(pPlayer->m_nPlayerID))
+        {
+            if (pPlayerInfo->m_bGameMasterPrivileges && !pPlayerInfo->m_bGameMasterIsPlayerLogin)
+            {
+                LOG_WARNING("ToggleDM: Called on a DMClient DM");
+                return Services::Events::Arguments();
+            }
+
+            if (auto *pMessage = Globals::AppManager()->m_pServerExoApp->GetNWSMessage())
+            {
+                bool currentlyPlayerDM = pPlayerInfo->m_bGameMasterPrivileges && pPlayerInfo->m_bGameMasterIsPlayerLogin;
+
+                if (isDM && !currentlyPlayerDM)
+                {
+                    pMessage->SendServerToPlayerDungeonMasterLoginState(pPlayer, true, true);
+                    pPlayerInfo->m_bGameMasterPrivileges = true;
+                    pPlayerInfo->m_bGameMasterIsPlayerLogin = true;
+                }
+                else if (!isDM && currentlyPlayerDM)
+                {
+                    pMessage->SendServerToPlayerDungeonMasterLoginState(pPlayer, false, true);
+                    pPlayerInfo->m_bGameMasterPrivileges = false;
+                    pPlayerInfo->m_bGameMasterIsPlayerLogin = false;
+                }
             }
         }
     }